refactor: Architect for multi-agent support
Transition from a single-purpose script to a scalable, multi-agent framework.

This refactoring introduces a robust architecture to support future development of various social media agents.

Key changes:
-   Introduced a `src` directory and a modern package structure with an installable `x-agent` entry point.
-   Isolated all X API interactions into a reusable `XService` class to centralize authentication, rate-limit handling, and API calls.
-   Created an abstract `BaseAgent` class to ensure a consistent interface for all future agents.
-   Refactored the original unblocking logic into a dedicated `UnblockAgent` that inherits from `BaseAgent`.
-   Implemented a central CLI entry point (`x_agent.cli`) to select and run agents via command-line arguments.
