+++ pyproject.toml
@@ -1,10 +1,16 @@
 [project]
-name = "x-unblocker"
+name = "x-agent"
 version = "0.1.0"
-description = "A command-line tool to unblock all blocked accounts on X."
+description = "A command-line tool to manage X interactions with agents."
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
     "python-dotenv",
     "tweepy",
 ]
+
+[project.scripts]
+x-agent = "x_agent.cli:main"
+
+[tool.setuptools.packages.find]
+where = ["src"]
